+++ CHANGELOG.md
@@ -7,6 +7,7 @@ This project adheres to [Semantic Versioning](https://semver.org/).
 
 ### Added
 - Add support for PHP 8.3 (#1218)
+- Add CSS selectors exclusion feature (#1202)
 
 ### Changed
 - Disable HTML formatting by default (#1214)